fixes gitRepo path variable and adds validation to saveState functions [ci skip]

+++ pipeline/scriptFunctions/processState.sh
@@ -6,7 +6,11 @@ save_state_variables() {
   STATE_SAVE_LOCATION=/build/state/
   # process each parameter passed in
   for var in "$@"; do
-    echo -e "export "$var"="${!var} >> $STATE_SAVE_LOCATION/variable_state.env
+    if [[ ! -z $var ]]; then
+      echo -e "export "$var"="${!var} >> $STATE_SAVE_LOCATION/variable_state.env
+    else
+      echo "variable not found or is empty...state not saved"
+    fi
   done
   echo "variables saved to state successfully"
 }
@@ -15,7 +19,11 @@ save_state_files() {
   STATE_SAVE_LOCATION=/build/state/
   # process each parameter passed in
   for file in "$@"; do
-    cp $file $STATE_SAVE_LOCATION
+    if [[ -f $file ]]; then
+      cp $file $STATE_SAVE_LOCATION
+    else
+      echo "file requested is not found in location specified...state not saved"
+    fi
   done
   echo "files saved to state successfully"
 }
